feat: 4.21*: Blog List Expansion, step 9

+++ controllers/blogs.js
@@ -3,14 +3,7 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const { notFound } = require('../utils/middleware')
-
-// const getTokenFrom = (request) => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
+const { InvalidTokenError, UnauthorizedError } = require('../utils/error')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
@@ -37,7 +30,7 @@ blogsRouter.post('/', async (request, response) => {
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'invalid token' })
+    throw InvalidTokenError()
   }
 
   const user = await User.findById(decodedToken.id)
@@ -59,6 +52,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    throw InvalidTokenError()
+  }
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() != decodedToken.id) {
+    throw UnauthorizedError()
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   return response.status(204).end()
 })
